+++ test/Algebra/Graph/Test/Generic.hs
@@ -11,12 +11,12 @@
 -----------------------------------------------------------------------------
 module Algebra.Graph.Test.Generic (
     -- * Generic tests
-    Testsuite, testsuite, HTestsuite, hTestsuite, testShow, testFromAdjacencyList,
-    testFromAdjacencySets, testFromAdjacencyIntSets, testBasicPrimitives,
-    testToGraph, testIsSubgraphOf, testSize, testProperties, testAdjacencyList,
-    testPreSet, testPostSet, testPostIntSet, testGraphFamilies,
-    testTransformations, testDfsForest, testDfsForestFrom, testDfs, testTopSort,
-    testIsTopSort, testSplitVertex, testBind, testSimplify
+    Testsuite, testsuite, testShow, testFromAdjacencyList, testFromAdjacencySets,
+    testFromAdjacencyIntSets, testBasicPrimitives, testIsSubgraphOf, testSize,
+    testToGraph, testAdjacencyList, testPreSet, testPreIntSet, testPostSet,
+    testPostIntSet, testGraphFamilies, testTransformations, testDfsForest,
+    testDfsForestFrom, testDfs, testTopSort, testIsTopSort, testSplitVertex,
+    testBind, testSimplify
   ) where
 
 import Prelude ()
@@ -25,35 +25,26 @@ import Prelude.Compat
 import Control.Monad (when)
 import Data.Orphans ()
 
-import Data.Foldable (toList)
 import Data.List (nub)
 import Data.Tree
 import Data.Tuple
 
+import Algebra.Graph (Graph (..))
 import Algebra.Graph.Class (Graph (..))
+import Algebra.Graph.ToGraph (ToGraph (..))
 import Algebra.Graph.Test
 import Algebra.Graph.Test.API
-import Algebra.Graph.Relation (Relation)
 
 import qualified Data.Set    as Set
 import qualified Data.IntSet as IntSet
 
 data Testsuite where
-    Testsuite :: (Arbitrary g, Eq g, GraphAPI g, Num g, Show g, Vertex g ~ Int)
+    Testsuite :: (Arbitrary g, Eq g, GraphAPI g, Num g, Show g, ToGraph g, ToVertex g ~ Int, Vertex g ~ Int)
               => String -> (forall r. (g -> r) -> g -> r) -> Testsuite
 
-testsuite :: (Arbitrary g, Eq g, GraphAPI g, Num g, Show g, Vertex g ~ Int)
+testsuite :: (Arbitrary g, Eq g, GraphAPI g, Num g, Show g, ToGraph g, ToVertex g ~ Int, Vertex g ~ Int)
           => String -> g -> Testsuite
 testsuite prefix g = Testsuite prefix (\f x -> f (x `asTypeOf` g))
-
-data HTestsuite where
-    HTestsuite :: (Arbitrary g, Eq g, GraphAPI g, Num g, Show g, Vertex g ~ Int,
-                   g ~ f Int, Foldable f)
-               => String -> (forall r. (g -> r) -> g -> r) -> HTestsuite
-
-hTestsuite :: (Arbitrary g, Eq g, GraphAPI g, Num g, Show g, Vertex g ~ Int,
-               g ~ f Int, Foldable f) => String -> g -> HTestsuite
-hTestsuite prefix g = HTestsuite prefix (\f x -> f (x `asTypeOf` g))
 
 testBasicPrimitives :: Testsuite -> IO ()
 testBasicPrimitives = mconcat [ testEmpty
@@ -66,17 +57,24 @@ testBasicPrimitives = mconcat [ testEmpty
                               , testOverlays
                               , testConnects ]
 
-testProperties :: Testsuite -> IO ()
-testProperties = mconcat [ testIsEmpty
-                         , testHasVertex
-                         , testHasEdge
-                         , testVertexCount
-                         , testEdgeCount
-                         , testVertexList
-                         , testEdgeList
-                         , testVertexSet
-                         , testVertexIntSet
-                         , testEdgeSet ]
+testToGraph :: Testsuite -> IO ()
+testToGraph = mconcat [ testToGraphDefault
+                      , testFoldg
+                      , testIsEmpty
+                      , testHasVertex
+                      , testHasEdge
+                      , testVertexCount
+                      , testEdgeCount
+                      , testVertexList
+                      , testVertexSet
+                      , testVertexIntSet
+                      , testEdgeList
+                      , testEdgeSet
+                      , testAdjacencyList
+                      , testPreSet
+                      , testPreIntSet
+                      , testPostSet
+                      , testPostIntSet ]
 
 testGraphFamilies :: Testsuite -> IO ()
 testGraphFamilies = mconcat [ testPath
@@ -346,37 +344,6 @@ testFromAdjacencyIntSets (Testsuite prefix (%)) = do
     test "overlay (fromAdjacencyIntSets xs) (fromAdjacencyIntSets ys)       == fromAdjacencyIntSets (xs ++ ys)" $ \xs ys ->
           overlay (fromAdjacencyIntSets xs) % fromAdjacencyIntSets ys       == fromAdjacencyIntSets (xs ++ ys)
 
-testToGraph :: HTestsuite -> IO ()
-testToGraph (HTestsuite prefix (%)) = do
-    putStrLn $ "\n============ " ++ prefix ++ "toGraph ============"
-    test "      toGraph (g     :: Graph a  ) :: Graph a       == g" $ \g ->
-                toGraph % g                                   == g
-
-    test "show (toGraph (1 * 2 :: Graph Int) :: Relation Int) == \"edge 1 2\"" $
-          show (toGraph % (1 * 2)            :: Relation Int) == "edge 1 2"
-
-    test "\ntoGraph == foldg empty vertex overlay connect" $ \x ->
-          toGraph % x == id % foldg empty vertex overlay connect x
-
-    putStrLn $ "\n============ " ++ prefix ++ "foldg ============"
-    test "foldg empty vertex        overlay connect        == id" $ \x ->
-          foldg empty vertex        overlay connect x      == id % x
-
-    test "foldg empty vertex        overlay (flip connect) == transpose" $ \x ->
-          foldg empty vertex        overlay (flip connect)x== transpose % x
-
-    test "foldg []    return        (++)    (++)           == toList" $ \x ->
-          foldg []    return        (++)    (++) x         == toList % x
-
-    test "foldg 0     (const 1)     (+)     (+)            == length" $ \x ->
-          foldg 0     (const 1)     (+)     (+) x          == length % x
-
-    test "foldg 1     (const 1)     (+)     (+)            == size" $ \x ->
-          foldg 1     (const 1)     (+)     (+) x          == size % x
-
-    test "foldg True  (const False) (&&)    (&&)           == isEmpty" $ \x ->
-          foldg True  (const False) (&&)    (&&) x         == isEmpty % x
-
 testIsSubgraphOf :: Testsuite -> IO ()
 testIsSubgraphOf (Testsuite prefix (%)) = do
     putStrLn $ "\n============ " ++ prefix ++ "isSubgraphOf ============"
@@ -394,6 +361,90 @@ testIsSubgraphOf (Testsuite prefix (%)) = do
 
     test "isSubgraphOf (path xs)     (circuit xs)  == True" $ \xs ->
           isSubgraphOf (path xs)    % circuit xs   == True
+
+testToGraphDefault :: Testsuite -> IO ()
+testToGraphDefault (Testsuite prefix (%)) = do
+    putStrLn $ "\n============ " ++ prefix ++ "toGraph et al. ============"
+    test "toGraph                    == foldg Empty Vertex Overlay Connect" $ \x ->
+          toGraph % x                == foldg Empty Vertex Overlay Connect x
+
+    test "foldg                      == foldg . toGraph" $ \e (apply -> v) (applyFun2 -> o) (applyFun2 -> c) x ->
+          foldg e v o c x            == (foldg (e :: Int) v o c . toGraph) % x
+
+    test "isEmpty                    == foldg True (const False) (&&) (&&)" $ \x ->
+          isEmpty x                  == foldg True (const False) (&&) (&&) % x
+
+    test "size                       == foldg 1 (const 1) (+) (+)" $ \x ->
+          size x                     == foldg 1 (const 1) (+) (+) % x
+
+    test "hasVertex x                == foldg False (==x) (||) (||)" $ \x y ->
+          hasVertex x y              == foldg False (==x) (||) (||) % y
+
+    test "hasEdge x y                == elem (x,y) . edgeList" $ \x y z ->
+          hasEdge x y z              == (elem (x,y) . edgeList) % z
+
+    test "vertexCount                == Set.size . vertexSet" $ \x ->
+          vertexCount x              == (Set.size . vertexSet) % x
+
+    test "vertexList                 == Set.toAscList . vertexSet" $ \x ->
+          vertexList x               == (Set.toAscList . vertexSet) % x
+
+    test "vertexSet                  == foldg Set.empty Set.singleton Set.union Set.union" $ \x ->
+          vertexSet x                == foldg Set.empty Set.singleton Set.union Set.union % x
+
+    test "vertexIntSet               == foldg IntSet.empty IntSet.singleton IntSet.union IntSet.union" $ \x ->
+          vertexIntSet x             == foldg IntSet.empty IntSet.singleton IntSet.union IntSet.union % x
+
+    test "edgeCount                  == length . edgeList" $ \x ->
+          edgeCount x                == (length . edgeList) % x
+
+    test "edgeList                   == edgeList . toAdjacencyMap" $ \x ->
+          edgeList x                 == (edgeList . toAdjacencyMap) % x
+
+    test "edgeSet                    == edgeSet . toAdjacencyMap" $ \x ->
+          edgeSet x                  == (edgeSet . toAdjacencyMap) % x
+
+    test "adjacencyList              == adjacencyList . toAdjacencyMap" $ \x ->
+          adjacencyList x            == (adjacencyList . toAdjacencyMap) % x
+
+    test "preSet x                   == postSet x . toAdjacencyMapTranspose" $ \x y ->
+          preSet x y                 == (postSet x . toAdjacencyMapTranspose) % y
+
+    test "preIntSet x                == IntSet.fromAscList . Set.toAscList . preSet x" $ \x y ->
+          preIntSet x y              == (IntSet.fromAscList . Set.toAscList . preSet x) % y
+
+    test "postSet x                  == postSet x . toAdjacencyMap" $ \x y ->
+          postSet x y                == (postSet x . toAdjacencyMap) % y
+
+    test "postIntSet x               == IntSet.fromAscList . Set.toAscList . postSet x" $ \x y ->
+          postIntSet x y             == (IntSet.fromAscList . Set.toAscList . postSet x) % y
+
+    test "toAdjacencyMap             == foldg empty vertex overlay connect" $ \x ->
+          toAdjacencyMap x           == foldg empty vertex overlay connect % x
+
+    test "toAdjacencyMapTranspose    == foldg empty vertex overlay (flip connect)" $ \x ->
+          toAdjacencyMapTranspose x  == foldg empty vertex overlay (flip connect) % x
+
+    test "toIntAdjacencyMap          == foldg empty vertex overlay connect" $ \x ->
+          toIntAdjacencyMap x        == foldg empty vertex overlay connect % x
+
+    test "toIntAdjacencyMapTranspose == foldg empty vertex overlay (flip connect)" $ \x ->
+          toIntAdjacencyMapTranspose x == foldg empty vertex overlay (flip connect) % x
+
+testFoldg :: Testsuite -> IO ()
+testFoldg (Testsuite prefix (%)) = do
+    putStrLn $ "\n============ " ++ prefix ++ "foldg ============"
+    test "foldg empty vertex        overlay connect        == id" $ \x ->
+          foldg empty vertex        overlay connect % x    == id x
+
+    test "foldg empty vertex        overlay (flip connect) == transpose" $ \x ->
+          foldg empty vertex        overlay (flip connect) % x == transpose x
+
+    test "foldg 1     (const 1)     (+)     (+)            == size" $ \x ->
+          foldg 1     (const 1)     (+)     (+) % x        == size x
+
+    test "foldg True  (const False) (&&)    (&&)           == isEmpty" $ \x ->
+          foldg True  (const False) (&&)    (&&) % x       == isEmpty x
 
 testIsEmpty :: Testsuite -> IO ()
 testIsEmpty (Testsuite prefix (%)) = do
@@ -615,6 +666,21 @@ testPostSet (Testsuite prefix (%)) = do
     test "postSet 2 (edge 1 2) == Set.empty" $
           postSet 2 % edge 1 2 == Set.empty
 
+testPreIntSet :: Testsuite -> IO ()
+testPreIntSet (Testsuite prefix (%)) = do
+    putStrLn $ "\n============ " ++ prefix ++ "preIntSet ============"
+    test "preIntSet x empty      == IntSet.empty" $ \x ->
+          preIntSet x % empty    == IntSet.empty
+
+    test "preIntSet x (vertex x) == IntSet.empty" $ \x ->
+          preIntSet x % vertex x == IntSet.empty
+
+    test "preIntSet 1 (edge 1 2) == IntSet.empty" $
+          preIntSet 1 % edge 1 2 == IntSet.empty
+
+    test "preIntSet y (edge x y) == IntSet.fromList [x]" $ \x y ->
+          preIntSet y % edge x y == IntSet.fromList [x]
+
 testPostIntSet :: Testsuite -> IO ()
 testPostIntSet (Testsuite prefix (%)) = do
     putStrLn $ "\n============ " ++ prefix ++ "postIntSet ============"
@@ -624,11 +690,11 @@ testPostIntSet (Testsuite prefix (%)) = do
     test "postIntSet x (vertex x) == IntSet.empty" $ \x ->
           postIntSet x % vertex x == IntSet.empty
 
-    test "postIntSet x (edge x y) == IntSet.fromList [y]" $ \x y ->
-          postIntSet x % edge x y == IntSet.fromList [y]
-
     test "postIntSet 2 (edge 1 2) == IntSet.empty" $
           postIntSet 2 % edge 1 2 == IntSet.empty
+
+    test "postIntSet x (edge x y) == IntSet.fromList [y]" $ \x y ->
+          postIntSet x % edge x y == IntSet.fromList [y]
 
 testPath :: Testsuite -> IO ()
 testPath (Testsuite prefix (%)) = do